Refactor: 기존 common 패키지 이름 변경 -> global
