+++ src/main/java/com/toty/following/presentation/FollowController.java
@@ -1,6 +1,6 @@
 package com.toty.following.presentation;
 
-import com.toty.annotation.CurrentUser;
+import com.toty.global.annotation.CurrentUser;
 import com.toty.following.application.FollowService;
 import com.toty.following.dto.request.FollowingRequest;
 import com.toty.following.dto.response.FollowingListResponse;
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RequestMapping("/api/follow")
-@Controller
+@RestController
 @RequiredArgsConstructor
 public class FollowController {
 
@@ -21,32 +21,32 @@ public class FollowController {
 
     // 팔로우하기
     @PostMapping("/")
-    @ResponseBody
-    public ResponseEntity follow (@CurrentUser User user, @RequestBody FollowingRequest followingRequest) {
+    public ResponseEntity<Long> follow(@CurrentUser User user,
+                                       @RequestBody FollowingRequest followingRequest) {
         Long response = followService.follow(user.getId(), followingRequest.getId());
-        return new ResponseEntity<>(response, HttpStatus.OK);
+        return ResponseEntity.ok(response);
     }
 
     // 언팔로우하기
     @DeleteMapping("/{id}")
-    @ResponseBody
-    public ResponseEntity unfollow (@CurrentUser User user, @PathParam("id") Long id) {
+    public ResponseEntity<Long> unfollow(@CurrentUser User user, @PathParam("id") Long id) {
         Long response = followService.unfollow(user.getId(), id);
-        return new ResponseEntity<>(response, HttpStatus.OK);
+        return ResponseEntity.ok(response);
     }
 
     // 팔로워 목록 조회
     @GetMapping("/{uid}/followers")
-    public ResponseEntity<FollowingListResponse> followersList(@PathParam("uid") Long id, @RequestParam(value = "p", defaultValue = "1") int page) {
+    public ResponseEntity<FollowingListResponse> followersList(@PathParam("uid") Long id,
+                                                               @RequestParam(value = "p", defaultValue = "1") int page) {
         FollowingListResponse response = followService.pagedFollowings(id, true, page);
         return ResponseEntity.ok(response);
     }
 
     // 팔로잉 목록 조회
     @GetMapping("/{uid}/followings")
-    public ResponseEntity<FollowingListResponse> followingList(@PathParam("uid") Long id, @RequestParam(value = "p", defaultValue = "1") int page) {
+    public ResponseEntity<FollowingListResponse> followingList(@PathParam("uid") Long id,
+                                                               @RequestParam(value = "p", defaultValue = "1") int page) {
         FollowingListResponse response = followService.pagedFollowings(id, false, page);
         return ResponseEntity.ok(response);
     }
-
 }
