+++ src/main/java/com/toty/global/annotation/CurrentUser.java
@@ -1,4 +1,4 @@
-package com.toty.annotation;
+package com.toty.global.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
