+++ src/main/java/com/toty/user/domain/repository/UserTagRepository.java
@@ -6,9 +6,10 @@ import com.toty.user.domain.model.UserTag;
 import org.springframework.data.repository.Repository;
 
 public interface UserTagRepository extends Repository<UserTag, Long> {
+
     List<UserTag> findByUserId(Long userId);
 
     void deleteByUserId(Long userId);
 
-    UserTag save(UserTag userTag);
+    void save(UserTag userTag);
 }
